Struct merkletree::merkle::MerkleTree

pub struct MerkleTree<E, A, S, BaseTreeArity = U2, SubTreeArity = U0, TopTreeArity = U0>where
    E: Element,
    A: Algorithm<E>,
    S: Store<E>,
    BaseTreeArity: Unsigned,
    SubTreeArity: Unsigned,
    TopTreeArity: Unsigned,{ /* private fields */ }

Implementations

impl<E: Element, A: Algorithm<E>, BaseTreeArity: Unsigned, SubTreeArity: Unsigned, TopTreeArity: Unsigned> MerkleTree<E, A, LevelCacheStore<E, File>, BaseTreeArity, SubTreeArity, TopTreeArity>

pub fn set_external_reader_path(&mut self, path: &Path) -> Result<()>

Given a pathbuf, instantiate an ExternalReader and set it for the LevelCacheStore.

pub fn from_store_configs_and_replica( leafs: usize, configs: &[StoreConfig], replica_config: &ReplicaConfig ) -> Result<MerkleTree<E, A, LevelCacheStore<E, File>, BaseTreeArity, SubTreeArity, TopTreeArity>>

Given a set of StoreConfig’s (i.e on-disk references to levelcache stores) and replica config info, instantiate each tree and return a compound merkle tree with them. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn from_sub_tree_store_configs_and_replica( leafs: usize, configs: &[StoreConfig], replica_config: &ReplicaConfig ) -> Result<MerkleTree<E, A, LevelCacheStore<E, File>, BaseTreeArity, SubTreeArity, TopTreeArity>>

Given a set of StoreConfig’s (i.e on-disk references to levelcache stores) and replica config info, instantiate each sub tree and return a compound merkle tree with them. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

impl<E: Element, A: Algorithm<E>, S: Store<E>, BaseTreeArity: Unsigned, SubTreeArity: Unsigned, TopTreeArity: Unsigned> MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>

pub fn new<I: IntoIterator<Item = E>>( data: I ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Creates new merkle from a sequence of hashes.

pub fn new_with_config<I: IntoIterator<Item = E>>( data: I, config: StoreConfig ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Creates new merkle from a sequence of hashes.

pub fn from_data<O: Hashable<A>, I: IntoIterator<Item = O>>( data: I ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Creates new merkle tree from a list of hashable objects.

pub fn from_data_with_config<O: Hashable<A>, I: IntoIterator<Item = O>>( data: I, config: StoreConfig ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Creates new merkle tree from a list of hashable objects.

pub fn from_data_store( data: S, leafs: usize ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Creates new merkle tree from an already allocated ‘Store’ (used with ‘Store::new_from_disk’). The specified ‘size’ is the number of base data leafs in the MT.

pub fn from_tree_slice( data: &[u8], leafs: usize ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Represent a fully constructed merkle tree from a provided slice.

pub fn from_tree_slice_with_config( data: &[u8], leafs: usize, config: StoreConfig ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Represent a fully constructed merkle tree from a provided slice.

pub fn from_trees( trees: Vec<MerkleTree<E, A, S, BaseTreeArity>> ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Creates new compound merkle tree from a vector of merkle trees. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn from_sub_trees( trees: Vec<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity>> ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Creates new top layer merkle tree from a vector of merkle trees with sub-trees. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn from_sub_trees_as_trees( trees: Vec<MerkleTree<E, A, S, BaseTreeArity>> ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Creates new top layer merkle tree from a vector of merkle trees by first constructing the appropriate sub-trees. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn from_slices( tree_data: &[&[u8]], leafs: usize ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity>>

Create a compound merkle tree given already constructed merkle trees contained as a slices. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn from_slices_with_configs( tree_data: &[&[u8]], leafs: usize, configs: &[StoreConfig] ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Create a compound merkle tree given already constructed merkle trees contained as a slices, along with configs for persistence. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn from_stores( leafs: usize, stores: Vec<S> ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Given a set of Stores (i.e. backing to MTs), instantiate each tree and return a compound merkle tree with them. The ordering of the stores is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn from_store_configs( leafs: usize, configs: &[StoreConfig] ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Given a set of StoreConfig’s (i.e on-disk references to disk stores), instantiate each tree and return a compound merkle tree with them. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn from_sub_tree_store_configs( leafs: usize, configs: &[StoreConfig] ) -> Result<MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>>

Given a set of StoreConfig’s (i.e on-disk references to dis stores), instantiate each sub tree and return a compound merkle tree with them. The ordering of the trees is significant, as trees are leaf indexed / addressable in the same sequence that they are provided here.

pub fn gen_proof(&self, i: usize) -> Result<Proof<E, BaseTreeArity>>

Generate merkle tree inclusion proof for leaf i

pub fn gen_cached_proof( &self, i: usize, rows_to_discard: Option<usize> ) -> Result<Proof<E, BaseTreeArity>>

Generate merkle tree inclusion proof for leaf i by first building a partial tree (returned) along with the proof. ‘rows_to_discard’ is an option that will be used if set (even if it may cause an error), otherwise a reasonable default is chosen.

Return value is a Result tuple of the proof and the partial tree that was constructed.

pub fn root(&self) -> E

Returns merkle root

pub fn len(&self) -> usize

Returns number of elements in the tree.

pub fn compact( &mut self, config: StoreConfig, store_version: u32 ) -> Result<bool>

Truncates the data for later access via LevelCacheStore interface.

pub fn reinit(&mut self) -> Result<()>

pub fn delete(&self, config: StoreConfig) -> Result<()>

Removes the backing store for this merkle tree.

pub fn is_empty(&self) -> bool

Returns true if the store contains no elements.

pub fn row_count(&self) -> usize

Returns row_count of the tree

pub fn leafs(&self) -> usize

Returns original number of elements the tree was built upon.

pub fn data(&self) -> Option<&S>

Returns data reference

pub fn read_at(&self, i: usize) -> Result<E>

Returns merkle leaf at index i

pub fn read_range(&self, start: usize, end: usize) -> Result<Vec<E>>

pub fn read_range_into( &self, start: usize, end: usize, buf: &mut [u8] ) -> Result<()>

pub fn read_into(&self, pos: usize, buf: &mut [u8]) -> Result<()>

Reads into a pre-allocated slice (for optimization purposes).

pub fn from_byte_slice_with_config( leafs: &[u8], config: StoreConfig ) -> Result<Self>

Build the tree given a slice of all leafs, in bytes form.

pub fn from_byte_slice(leafs: &[u8]) -> Result<Self>

Build the tree given a slice of all leafs, in bytes form.

impl<E: Element, A: Algorithm<E>, S: Store<E>, BaseTreeArity: Unsigned, SubTreeArity: Unsigned, TopTreeArity: Unsigned> MerkleTree<E, A, S, BaseTreeArity, SubTreeArity, TopTreeArity>

pub fn try_from_iter<I: IntoIterator<Item = Result<E>>>(into: I) -> Result<Self>

Attempts to create a new merkle tree using hashable objects yielded by the provided iterator. This method returns the first error yielded by the iterator, if the iterator yielded an error.

pub fn try_from_iter_with_config<I: IntoIterator<Item = Result<E>>>( into: I, config: StoreConfig ) -> Result<Self>

Attempts to create a new merkle tree using hashable objects yielded by the provided iterator and store config. This method returns the first error yielded by the iterator, if the iterator yielded an error.