Struct merkletree::store::LevelCacheStore

pub struct LevelCacheStore<E: Element, R: Read + Send + Sync> { /* private fields */ }

The LevelCacheStore is used to reduce the on-disk footprint even further to the minimum at the cost of build time performance. Each LevelCacheStore is created with a StoreConfig object which contains the number of binary tree levels above the base that are ‘cached’. This implementation has hard requirements about the on disk file size based on that number of levels, so on-disk files are tied, structurally to the configuration they were built with and can only be accessed with the same number of levels.

Implementations

impl<E: Element, R: Read + Send + Sync> LevelCacheStore<E, R>

pub fn new_from_disk_with_reader( store_range: usize, branches: usize, config: &StoreConfig, reader: ExternalReader<R> ) -> Result<Self>

Used for opening v2 compacted DiskStores.

pub fn set_external_reader(&mut self, reader: ExternalReader<R>) -> Result<()>

impl<E: Element, R: Read + Send + Sync> LevelCacheStore<E, R>

pub fn set_len(&mut self, len: usize)

pub fn front_truncate(&mut self, config: &StoreConfig, len: usize) -> Result<()>

pub fn store_size(&self) -> usize

pub fn is_consistent_v1( store_range: usize, branches: usize, config: &StoreConfig ) -> Result<bool>

pub fn is_consistent( store_range: usize, branches: usize, config: &StoreConfig ) -> Result<bool>

pub fn store_read_range(&self, start: usize, end: usize) -> Result<Vec<u8>>

pub fn store_read_into( &self, start: usize, end: usize, buf: &mut [u8] ) -> Result<()>

pub fn store_copy_from_slice( &mut self, start: usize, slice: &[u8] ) -> Result<()>